Html.Attributes.value docs should mention textarea
I was mislead by the docs for the `Html.Attributes.value` function, since it didn't mention that it worked with `textarea`. I was instead trying to bind the contents of the `textarea` to the underlying model by setting the value of an inner text node.

Thanks for elm! It's been really great so far!